Whole tool folder/file structure refactored
