+++ internal/aggregator/interfaces/IAggregator.go
