+++ internal/configuration/interfaces/IConfiguration.go
