+++ internal/ping/interfaces/IPingService.go
@@ -4,4 +4,5 @@ type IPingService interface {
 	String() string
 	Configure()
 	StartMethods()
+	GetMethods() []IPingMethod
 }
