+++ internal/processor/availability/PercentageProcessedData.go
