+++ internal/processor/availability/PercentageProcessor.go
