+++ internal/processor/availability/UnitProcessedData.go
