+++ internal/processor/types/ProcessorParams.go
